requirements.txt
redis==3.5.3
gunicorn==20.1.0
python-dotenv==0.18.0

flask==1.1.4    #20210602測試2.0.1匯出pmenu不穩定
flask_cors==3.0.10
mysqlclient==2.0.3
sqlalchemy==1.4.15

